+++ src/main/index.ts
@@ -1,55 +1,11 @@
-const ERR_NOT_SUPPORTED = 'Not Supported';
+import {
+  ERR_NOT_SUPPORTED,
+  CHARACTER_SET_SIMPLIFIED_NORMAL,
+  CHARACTER_SET_SIMPLIFIED_CAPITAL,
+  CHARACTER_SET_TRADITIONAL_NORMAL,
+  CHARACTER_SET_TRADITIONAL_CAPITAL,
+} from './Constants';
 
-const CHARACTER_SET_SIMPLIFIED_NORMAL = {
-  BASE: '零一二三四五六七八九',
-  ZERO: '零',
-  TEN: '十',
-  HUNDRED: '百',
-  THOUSAND: '千',
-  TENTHOUSAND: '万',
-  HUNDREDMILLION: '亿',
-  TRILLION: '兆',
-  NEGATIVE: '负',
-  POSITIVE: '正',
-};
-
-const CHARACTER_SET_SIMPLIFIED_CAPITAL = {
-  BASE: '零壹贰叁肆伍陆柒捌玖',
-  ZERO: '零',
-  TEN: '拾',
-  HUNDRED: '佰',
-  THOUSAND: '仟',
-  TENTHOUSAND: '万',
-  HUNDREDMILLION: '亿',
-  TRILLION: '兆',
-  NEGATIVE: '负',
-  POSITIVE: '正',
-};
-
-const CHARACTER_SET_TRADITIONAL_NORMAL = {
-  BASE: '零一二三四五六七八九',
-  ZERO: '零',
-  TEN: '十',
-  HUNDRED: '百',
-  THOUSAND: '千',
-  TENTHOUSAND: '萬',
-  HUNDREDMILLION: '億',
-  TRILLION: '兆',
-  NEGATIVE: '負',
-  POSITIVE: '正',
-};
-const CHARACTER_SET_TRADITIONAL_CAPITAL = {
-  BASE: '零壹貳參肆伍陸柒捌玖',
-  ZERO: '零',
-  TEN: '拾',
-  HUNDRED: '佰',
-  THOUSAND: '仟',
-  TENTHOUSAND: '萬',
-  HUNDREDMILLION: '億',
-  TRILLION: '兆',
-  NEGATIVE: '負',
-  POSITIVE: '正',
-};
 class Convertor {
   characterSet: typeof CHARACTER_SET_SIMPLIFIED_NORMAL;
   constructor(options: options) {