+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## `1.5.1` - 17 June 2022
 
 - `#29`, As user, I want to translate into capital numbers (大写数字) - traditional
-- fix bugs
-- refactor test for `0`, `1-9`
+- Fix bugs
+- Refactor test for `0`, `1-9`
+- Refactor into `Constants.ts`
 
 ## `1.5.0` - 16 June 2022